Extract the agent-primary launch picker into a shared control
New, Resume, and Schedule should render an identical Agent+Transport
control; pull AgentPicker and TransportPicker out of LaunchPanel /
LaunchOptions into a reusable AgentTransportPicker, with a
useTransportForAgent hook that owns the transport state and re-clamps it
to the selected agent's supported set. LaunchPanel now composes them and
behaves unchanged.
