+++ frontend/src/components/LaunchOptions.tsx
@@ -2,9 +2,7 @@
 
 import { useState } from "react";
 
-import type { BackendCatalog } from "@/lib/backends";
-import { transportPresentation } from "@/lib/backends";
-import { LaunchMode, SessionTransport } from "@/lib/types";
+import { LaunchMode } from "@/lib/types";
 
 export function GearGlyph() {
   return (
@@ -191,139 +189,5 @@ export function LaunchModeField({
         ))}
       </div>
     </label>
-  );
-}
-
-interface TransportPickerProps {
-  transports: SessionTransport[];
-  value: SessionTransport;
-  onChange: (transport: SessionTransport) => void;
-  catalog: BackendCatalog;
-}
-
-// Agent-primary transport selector. Populated from an agent's
-// supported_transports and rendered as a light segmented control: each
-// transport gets a distinct icon and its user-facing name, with the selected
-// transport's one-line description shown beneath. Collapses to nothing when the
-// agent exposes a single transport, since there is nothing to choose.
-export function TransportPicker({
-  transports,
-  value,
-  onChange,
-  catalog,
-}: TransportPickerProps) {
-  if (transports.length <= 1) {
-    return null;
-  }
-  const selected = transportPresentation(value, catalog);
-  return (
-    <div className="field transport-field">
-      <span>Interface</span>
-      <div
-        className="segmented segmented-quiet transport-segmented"
-        role="radiogroup"
-        aria-label="Interface"
-      >
-        {transports.map((transport) => {
-          const { name, description, kind } = transportPresentation(
-            transport,
-            catalog,
-          );
-          const active = value === transport;
-          return (
-            <button
-              key={transport}
-              type="button"
-              role="radio"
-              aria-checked={active}
-              title={description}
-              className={`segmented-item transport-segment${active ? " active" : ""}`}
-              onClick={() => onChange(transport)}
-            >
-              <span className="transport-segment-icon" aria-hidden="true">
-                <TransportGlyph transport={transport} kind={kind} />
-              </span>
-              {name}
-            </button>
-          );
-        })}
-      </div>
-      <p className="transport-desc">{selected.description}</p>
-    </div>
-  );
-}
-
-// Distinct glyph per transport: a speech bubble for the structured Chat
-// adapter, an app window for the Emulated (real-app) tail, and a terminal
-// prompt for the raw Terminal pane. Falls back to the kind for unknown
-// transports.
-function TransportGlyph({
-  transport,
-  kind,
-}: {
-  transport: SessionTransport;
-  kind: "chat" | "terminal";
-}) {
-  if (transport === "claude_tty") return <EmulatedGlyph />;
-  if (kind === "terminal") return <TerminalGlyph />;
-  return <ChatGlyph />;
-}
-
-function ChatGlyph() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
-      <path
-        d="M3 2.9h10A1.3 1.3 0 0 1 14.3 4.2v4.9A1.3 1.3 0 0 1 13 10.4H6.7L4 12.8V10.4H3A1.3 1.3 0 0 1 1.7 9.1V4.2A1.3 1.3 0 0 1 3 2.9Z"
-        stroke="currentColor"
-        strokeWidth="1.2"
-        strokeLinejoin="round"
-      />
-      <path d="M4.3 5.4h7.4" stroke="currentColor" strokeWidth="1.2" strokeLinecap="round" opacity="0.65" />
-      <path d="M4.3 7.5h4.8" stroke="currentColor" strokeWidth="1.2" strokeLinecap="round" opacity="0.65" />
-    </svg>
-  );
-}
-
-function EmulatedGlyph() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
-      <rect
-        x="1.7"
-        y="2.7"
-        width="12.6"
-        height="10.6"
-        rx="1.7"
-        stroke="currentColor"
-        strokeWidth="1.2"
-      />
-      <path d="M1.7 5.5h12.6" stroke="currentColor" strokeWidth="1.2" opacity="0.65" />
-      <circle cx="3.8" cy="4.1" r="0.55" fill="currentColor" />
-      <circle cx="5.6" cy="4.1" r="0.55" fill="currentColor" opacity="0.6" />
-      <path d="M4.4 8h6.2M4.4 10.3h3.9" stroke="currentColor" strokeWidth="1.2" strokeLinecap="round" opacity="0.65" />
-    </svg>
-  );
-}
-
-function TerminalGlyph() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 16 16" fill="none" aria-hidden="true">
-      <rect
-        x="1.6"
-        y="2.6"
-        width="12.8"
-        height="10.8"
-        rx="1.6"
-        stroke="currentColor"
-        strokeWidth="1.2"
-      />
-      <path
-        d="M4.3 6.2l2 1.9-2 1.9"
-        stroke="currentColor"
-        strokeWidth="1.2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path d="M7.8 10.2h3.6" stroke="currentColor" strokeWidth="1.2" strokeLinecap="round" />
-    </svg>
   );
 }
